rename modules per idiomatic rust
Note there is no public change here - the modules are all private,
except for geodesic_capability, which continues to be exported via the
`capability` alias.
